+++ src/nativeOptions.ts
@@ -1,3 +1,5 @@
+import { Capacitor } from '@capacitor/core';
+
 import type { CapacitorOptions } from './options';
 
 /**
@@ -34,5 +36,14 @@ export function FilterNativeOptions(options: CapacitorOptions): CapacitorOptions
     tracesSampleRate: options.tracesSampleRate,
     // tunnel: options.tunnel: Only handled on the JavaScript Layer.
     enableCaptureFailedRequests: options.enableCaptureFailedRequests,
+    ...iOSParameters(options)
   };
+}
+
+function iOSParameters(options: CapacitorOptions): CapacitorOptions
+{
+  return Capacitor.getPlatform() === 'ios' ? {
+    enableAppHangTracking: options.enableAppHangTracking,
+    appHangTimeoutInterval: options.appHangTimeoutInterval
+  } : {}
 }